refactor the test base to be shared between sync and async
