+++ azure-quantum/tests/unit/common.py
@@ -9,10 +9,14 @@
 
 import os
 import re
+
 import six
+import pytest
 
 from azure.quantum import Workspace
-from azure.identity import DefaultAzureCredential, ClientSecretCredential
+from azure.identity import ClientSecretCredential
+from azure.quantum.aio import Workspace as AsyncWorkspace
+from azure.identity.aio import ClientSecretCredential as AsyncClientSecretCredential
 from azure_devtools.scenario_tests.base import ReplayableTest
 from azure_devtools.scenario_tests.recording_processors import (
     RecordingProcessor,
@@ -33,6 +37,7 @@ WORKSPACE = "myworkspace"
 LOCATION = "eastus"
 STORAGE = "mystorage"
 
+@pytest.mark.usefixtures("event_loop_instance")
 class QuantumTestBase(ReplayableTest):
     """QuantumTestBase
 
@@ -190,6 +195,34 @@ class QuantumTestBase(ReplayableTest):
                              else None
 
         workspace = Workspace(
+            credential=default_credential,
+            subscription_id=self.subscription_id,
+            resource_group=self.resource_group,
+            name=self.workspace_name,
+            location=self.location,
+        )
+
+        return workspace
+
+    def get_async_result(self, coro):
+        return self.event_loop.run_until_complete(coro)
+
+    def create_async_workspace(self) -> AsyncWorkspace:
+        """Create workspace using credentials passed via OS Environment Variables
+        described in the README.md documentation, or when in playback mode use
+        a placeholder credential.
+
+        :return: AsyncWorkspace
+        :rtype: AsyncWorkspace
+        """
+
+        playback_credential = AsyncClientSecretCredential(self.tenant_id,
+                                                     self.client_id,
+                                                     self.client_secret)
+        default_credential = playback_credential if self.is_playback \
+                             else None
+
+        workspace = AsyncWorkspace(
             credential=default_credential,
             subscription_id=self.subscription_id,
             resource_group=self.resource_group,
