+++ azure-quantum/tests/unit/test_job.py
@@ -13,7 +13,7 @@ import functools
 import pytest
 from datetime import date, datetime, timedelta
 
-from common import QuantumTestBase, ZERO_UID
+from .common import QuantumTestBase, ZERO_UID
 from azure.quantum import Job
 from azure.quantum.optimization import Problem, ProblemType, Term, GroupedTerm, GroupType
 import azure.quantum.optimization as microsoft
