+++ azure-quantum/tests/unit/test_optimization.py
@@ -19,7 +19,7 @@ from azure.quantum.optimization.solvers import (
     QuantumMonteCarlo,
     SubstochasticMonteCarlo,
 )
-from common import QuantumTestBase
+from .common import QuantumTestBase
 
 
 class TestProblem(QuantumTestBase):
