+++ azure-quantum/tests/unit/test_problem.py
@@ -12,7 +12,7 @@ from unittest.mock import Mock, patch
 from typing import TYPE_CHECKING
 from azure.quantum.optimization import Problem, ProblemType, Term, GroupedTerm
 import azure.quantum.optimization.problem
-from common import expected_terms
+from .common import expected_terms
 import numpy
 import os
