+++ azure-quantum/tests/unit/test_streaming_problem.py
@@ -20,7 +20,7 @@ from azure.quantum.optimization import (
     Term,
 )
 from azure.quantum.storage import download_blob
-from common import QuantumTestBase
+from .common import QuantumTestBase
 
 
 class TestStreamingProblem(QuantumTestBase):
