+++ azure-quantum/tests/unit/test_workspace.py
@@ -9,7 +9,7 @@
 import pytest
 
 from azure.quantum import Workspace
-from common import QuantumTestBase
+from .common import QuantumTestBase
 
 
 class TestWorkspace(QuantumTestBase):
